Add files via upload
In this code, the input function is used to ask the user to input a number, which is then passed as an argument to the is_prime function to determine whether it is a prime number. The output message is printed based on the result of the is_prime function.